api_bank_django/requirements/base.txt
django==4.2.15
python-dotenv==1.0.1
djangorestframework==3.15.2
django-countries==7.6.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.42
drf-spectacular==0.27
pillow==10.4.0
argon2-cffi==23.1.0
djoser==2.2.3
django-filter==24.3
django-celery-email==3.0.0
django-celery-beat==2.6.0
cloudinary==1.39.1
python-dateutil==2.9.0
psycopg2-binary==2.9.9